+++ template-lit-js/src/newtab/newtab.js
@@ -20,9 +20,9 @@ export class Newtab extends LitElement {
       <main>
         <h3>Newtab Page!</h3>
 
-        <h6>v 0.0.1</h6>
+        <h6>v 0.0.0</h6>
 
-        <a href="https://github.com/guocaoyi/create-chrome-ext" target="_blank"
+        <a href="https://www.npmjs.com/package/create-chrome-ext" target="_blank"
           >Generator by ${this.crx}</a
         >
       </main>