+++ template-lit-ts/types/manifest.d.ts
@@ -1,2 +1,0 @@
-declare const _default: import("@crxjs/vite-plugin").ManifestV3Export;
-export default _default;