+++ template-lit-ts/vite.config.ts
@@ -1,6 +1,7 @@
 import * as path from 'path'
 import { defineConfig } from 'vite'
 import { crx } from '@crxjs/vite-plugin'
+
 import manifest from './src/manifest'
 
 // https://vitejs.dev/config/