+++ template-preact-js/vite.config.js
@@ -1,6 +1,7 @@
 import { defineConfig } from 'vite'
-import preact from '@preact/preset-vite'
 import { crx } from '@crxjs/vite-plugin'
+import preact from '@preact/preset-vite'
+
 import manifest from './src/manifest.js'
 
 // https://vitejs.dev/config/