+++ template-react-js/src/options/Options.jsx
@@ -2,18 +2,15 @@ import { useState } from 'react'
 import './Options.css'
 
 function App() {
-  const [crx, setCrx] = useState('chrome-extension-boilerplate-vite-svelte')
+  const [crx, setCrx] = useState('create-chrome-ext')
 
   return (
     <main>
       <h3>Options Page!</h3>
 
-      <h6>v 0.0.1</h6>
+      <h6>v 0.0.0</h6>
 
-      <a
-        href="https://github.com/guocaoyi/chrome-extension-boilerplate-vite-svelte"
-        target="_blank"
-      >
+      <a href="https://www.npmjs.com/package/create-chrome-ext" target="_blank">
         Power by {crx}
       </a>
     </main>