+++ template-react-ts/vite.config.ts
@@ -1,6 +1,7 @@
 import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
 import { crx } from '@crxjs/vite-plugin'
+import react from '@vitejs/plugin-react'
+
 import manifest from './src/manifest'
 
 // https://vitejs.dev/config/