+++ template-svelte-js/vite.config.js
@@ -1,7 +1,7 @@
-import sveltePreprocess from 'svelte-preprocess'
 import { defineConfig } from 'vite'
 import { crx } from '@crxjs/vite-plugin'
 import { svelte } from '@sveltejs/vite-plugin-svelte'
+import sveltePreprocess from 'svelte-preprocess'
 
 import manifest from './src/manifest.js'