+++ template-svelte-ts/vite.config.ts
@@ -1,8 +1,8 @@
-import path from 'path'
-import sveltePreprocess from 'svelte-preprocess'
 import { defineConfig } from 'vite'
 import { crx } from '@crxjs/vite-plugin'
 import { svelte } from '@sveltejs/vite-plugin-svelte'
+import path from 'path'
+import sveltePreprocess from 'svelte-preprocess'
 
 import manifest from './src/manifest'