+++ template-vanilla-ts/src/popup/index.ts
@@ -6,10 +6,10 @@ document.querySelector<HTMLDivElement>('#app')!.innerHTML = `
   <main>
   <h3>Popup Page!</h3>
 
-  <h6>v 0.0.1</h6>
+  <h6>v 0.0.0</h6>
 
   <a
-    href="https://github.com/guocaoyi/create-chrome-ext"
+    href="https://www.npmjs.com/package/create-chrome-ext"
     target="_blank"
   >
     Power by ${crx}