+++ template-vanilla-ts/vite.config.ts
@@ -1,5 +1,6 @@
 import { defineConfig } from 'vite'
 import { crx } from '@crxjs/vite-plugin'
+
 import manifest from './src/manifest'
 
 // https://vitejs.dev/config/