+++ template-vue-ts/vite.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from 'vite'
-import vue from '@vitejs/plugin-vue'
 import { crx } from '@crxjs/vite-plugin'
+import vue from '@vitejs/plugin-vue'
 
 import manifest from './src/manifest'